requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
numpy==1.22.3
paho-mqtt==1.6.1
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
